Hammcco クライアントアプリを大きくリファクタリング
これでクライアントアプリは一応完成しました。
