+++ client/package.json
@@ -6,7 +6,8 @@
     "axios": "^0.16.1",
     "prop-types": "^15.5.10",
     "react": "^15.5.4",
-    "react-dom": "^15.5.4"
+    "react-dom": "^15.5.4",
+    "react-router-dom": "^4.1.1"
   },
   "devDependencies": {
     "eslint": "^3.19.0",
