+++ client/src/Hamcco.jsx
@@ -1,108 +1,64 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import './Hamcco.css';
+import Axios from 'axios';
+import HamccoPanel from './HamccoPanel';
+import HamccoInput from './HamccoInput';
+import HamccoGuide from './HamccoGuide';
 
 /**
- * Hamccoコンポーネントクラス
+ * Hamccoアプリケーション（メインコントローラ）クラス
  */
-function Hamcco(props) {
-  return (
-    <div className="hamcco-container">
-      <div className="hamcco-grid">
-        <HamccoImage feel={props.feel} flex="flex6" />
-        <HamccoMessage message={props.message} flex="flex6" />
-      </div>
-      <div className="hamcco-grid">
-        <HamccoLevel level={props.feel} />
-      </div>
-    </div>
-  );
-}
-Hamcco.propTypes = {
-  feel: PropTypes.number.isRequired,
-  message: PropTypes.string.isRequired,
-};
-
-/**
- * Hamcco イメージコンポーネントクラス
- */
-class HamccoImage extends Component {
-  getInfoByFeel() {
-    const dir = '/img/';
-    let info = { name: '天使ハム', image: `${dir}tensi_ham.gif` };
-    let flno = Math.floor(this.props.feel / 20) + 1;
-    flno = (flno > 5) ? 5 : flno;
-    flno = (flno < 1) ? 1 : flno;
-
-    switch (flno) {
-    case 1:   // Very BAD
-      info = { name: 'おこりハム', image: `${dir}okori_ham.gif` };
-      break;
-    case 2:   // BAD
-      info = { name: 'オニハム', image: `${dir}oni_ham.gif` };
-      break;
-    case 3:   // Normal
-      info = { name: 'キキハム', image: `${dir}kiki_ham.gif` };
-      break;
-    case 4:   // Good
-      info = { name: 'コックハム', image: `${dir}cook_ham.gif` };
-      break;
-    case 5:   // Very Good
-      info = { name: '天使ハム', image: `${dir}tensi_ham.gif` };
-      break;
-    default:
-      break;
-    }
-    return info;
+class Hamcco extends Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      feel: 80,
+      message: 'Hi! user',
+      count: 0,
+    };
   }
+
+  componentWillMount() {
+    this.setState(() => ({
+      message: `こんにちは ${this.props.username} さん、なにか話しかけてね！`,
+    }));
+  }
+
+  sendMessage(message) {
+    console.log(`say message: ${message}`);
+
+    Axios.post('/hamcco/chat', {
+      usrname: this.props.username,
+      count: this.state.count,
+      feel: this.state.feel,
+      message,
+    }).then((res) => {
+      console.log(`status : ${res.status}`);
+      console.log(`data : ${JSON.stringify(res.data)}`);
+
+      this.setState(() => ({
+        count: this.state.count + 1,
+        feel: res.data.feel,
+        message: res.data.message,
+      }));
+    }).catch((res) => {
+      console.log(res);
+    });
+  }
+
   render() {
-    const info = this.getInfoByFeel();
     return (
-      <div className={`hamcco-cell ${this.props.flex}`}>
-        <h4>{info.name}</h4>
-        <img src={info.image} alt={info.name} />
+      <div>
+        <HamccoPanel feel={this.state.feel} message={this.state.message} />
+        <HamccoInput onSubmit={m => this.sendMessage(m)} />
+        <HamccoGuide />
       </div>
     );
   }
 }
-HamccoImage.propTypes = {
-  feel: PropTypes.number.isRequired,
-  flex: PropTypes.string.isRequired,
-};
 
-/**
- * Hamccoメッセージコンポーネントクラス
- */
-function HamccoMessage(props) {
-  return (
-    <div className={`hamcco-cell ${props.flex}`}>
-      <div className="hamcco-message">
-        {props.message}
-      </div>
-    </div>
-  );
-}
-HamccoMessage.propTypes = {
-  flex: PropTypes.string.isRequired,
-  message: PropTypes.string.isRequired,
-};
-
-/**
- * Hamccoごきげんメータコンポーネントクラス
- */
-function HamccoLevel(props) {
-  return (
-    <div className="hamcco-cell">
-      ごきげんメータ：
-      <span>{props.level} %</span>
-      <meter min="0" max="100" low="20" high="40" optimum="99" value={props.level}>
-        {props.level} %
-      </meter>
-    </div>
-  );
-}
-HamccoLevel.propTypes = {
-  level: PropTypes.number.isRequired,
+Hamcco.propTypes = {
+  username: PropTypes.string.isRequired,
 };
 
 export default Hamcco;
