+++ client/src/HamccoGuide.css
@@ -1,4 +1,4 @@
-.hamguide-container {
+.hamcco-guide {
   width: 80%;
   max-width: 800px;
   min-width: 400px;
@@ -9,7 +9,7 @@
   text-align: left;
 }
 
-.hamguide-container section {
+.hamcco-guide section {
   margin: 0;
   padding: 5px 30px;
 }
