+++ client/src/HamccoPanel.css
@@ -1,4 +1,4 @@
-.hamcco-container {
+.hamcco-panel {
   width: 80%;
   max-width: 800px;
   min-width: 400px;
@@ -41,11 +41,12 @@
   box-sizing: border-box;
   height: 240px;
   margin: 40px 20px 20px 20px;
-  padding: 20px 10px;
+  padding: 50px 30px;
   border-radius: 20px;
   color: #FFFFFF;
   background-color: #66CC99;
   position: relative;
+  text-align: left;
 }
 .hamcco-message:after {
   border: 40px solid transparent;
@@ -58,4 +59,3 @@
   position: absolute;
   width: 0;
 }
-
