+++ client/src/index.css
@@ -3,20 +3,3 @@ body {
   padding: 0;
   font-family: sans-serif;
 }
-
-.App {
-  text-align: center;
-  margin: 0 auto;
-  padding: 20px;
-  background-color: #336699;
-}
-
-header {
-  padding: 5px;
-  color: white;
-}
-
-footer {
-  padding: 5px;
-  color: white;
-}
