+++ client/src/index.js
@@ -1,45 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import HamController from './HamController';
+import { BrowserRouter } from 'react-router-dom';
+import AppFrame from './AppFrame';
 import './index.css';
 
-ReactDOM.render(
-  <App />,
+ReactDOM.render((
+  <BrowserRouter>
+    <AppFrame />
+  </BrowserRouter>
+  ),
   document.getElementById('root'),
 );
-
-/**
- * Hamcco画面フレーム
- */
-function App() {
-  return (
-    <div className="App">
-      <HamHeader />
-      <HamController />
-      <HamFooter />
-    </div>
-  );
-}
-
-/**
- * Hamcco画面共通ヘッダ
- */
-function HamHeader() {
-  return (
-    <header>
-      <h1>人口無能ハムっこであそぼう</h1>
-    </header>
-  );
-}
-
-/**
- * Hamcco画面共通フッタ
- */
-function HamFooter() {
-  return (
-    <footer>
-      <small>French Connection</small>
-    </footer>
-  );
-}
-
